backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
azure-storage-blob==12.25.1
python-multipart==0.0.20
pydantic==2.11.5
pydantic-settings==2.2.1
python-magic==0.4.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1 